change ints to floats

+++ internal/app/internal/orderer/orderer.go
@@ -10,7 +10,7 @@ type Orderer interface {
 	NewOrder(context.Context, string, string) error
 	Orders(context.Context, string) ([]model.Order, error)
 	Balance(context.Context, string) (model.Balance, error)
-	Withdraw(context.Context, string, string, int) error
+	Withdraw(context.Context, string, string, float32) error
 	Withdrawals(context.Context, string) ([]model.Withdrawals, error)
 	OrdersChan() <-chan string
 }
